+++ src/components/index.js
@@ -1,8 +1,0 @@
-export { Counter }        from './Counter.js'
-export { ErrorCard }      from './ErrorCard.js'
-export { Link }           from './Link.js'
-export { LoadingCard }    from './LoadingCard.js'
-export { ThemeSelector }  from './ThemeSelector.js'
-export { ThemeToggle }    from './ThemeToggle.js'
-export { ToastContainer } from './ToastContainer.js'
-export { UserCard }       from './UserCard.js'